backend/requirements.txt
aiohttp==3.7.4.post0
fastapi==0.104.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
starlette==0.27.0
uvicorn==0.24.0.post1
pymongo==4.6.0
